requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
openai==1.30.1
anthropic==0.26.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
yookassa==2.3.0
python-multipart==0.0.9
